Ensure OpenStack resource are immutable
actuator.Update() could in certain circumstances cause a server to be
recreated. This is always an error and can only result in data loss.

This change removes all functionality from CAPO which relates to
recreating OpenStack resources. Instead, when we get an Update we simply
synchronise the Machine object with the current state of the OpenStack
resource.

Additionally, we are no longer setting the instance-status annotation on
the machine object, as this is no longer relevant. Instead we detect a
deleted OpenStack resource by setting the providerID on the machine.
This also has the advantage of making nodelink matching more robust.