Bug 2073398: Fix InstanceCreate port & trunk cleanup
With this change, the port and trunk cleanups are triggered by any
early failure of InstanceCreate.

When InstanceCreate creates ports and if it creates trunks, these
resources will be deleted if any of the next steps fail and cause the
server never to be created.

Note: the trunk will be deleted first, then the ports.