Remove dependence on annotation to allow Server deletion
If a failure happened while handling the Machine creation,
the Machine can end up without an openstack-resourceId
annotation. This makes the deletion of the Machine to fail as
no Server ID was specified in the annotation. This commit fixes
the issue by making the deletion rely on the Server ID retrieved
prior to delete instead of using the annotation.

Co-authored-by: Matthew Booth mbooth@redhat.com