OCPBUGS-2508: Ensure network defs without subnet follow noAllowedAddressPairs
Commit 855a22c09317457ecfd08544afac6948cfdfe462 introduced a bug that
caused network definitions without subnets (the one you would get when
setting `additionalNetworkIDs` in your `install-config.yaml`) caused the
noAllowedAddressPairs property to be skipped while creating the port.

This may in turn caused machine-api to fail creating the port because it
does not have the permission to create a port with allowed_address_pair,
which is often the case when using provider networks.